feat(frontend): split code generation in separate packages
