+++ frontend/flup_openapi_lib/.openapi-generator/VERSION
