+++ frontend/flup_openapi_lib/README.md
@@ -1,4 +1,4 @@
-# test_api
+# flup_openapi_lib
 No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 
 This Dart package is automatically generated by the [OpenAPI Generator](https://openapi-generator.tech) project:
@@ -16,7 +16,7 @@ Dart 2.12 or later
 If this Dart package is published to Github, add the following dependency to your pubspec.yaml
 ```
 dependencies:
-  test_api:
+  flup_openapi_lib:
     git: https://github.com/GIT_USER_ID/GIT_REPO_ID.git
 ```
 
@@ -24,8 +24,8 @@ dependencies:
 To use the package in your local drive, add the following dependency to your pubspec.yaml
 ```
 dependencies:
-  test_api:
-    path: /path/to/test_api
+  flup_openapi_lib:
+    path: /path/to/flup_openapi_lib
 ```
 
 ## Tests
@@ -37,7 +37,7 @@ TODO
 Please follow the [installation procedure](#installation--usage) and then run the following:
 
 ```dart
-import 'package:test_api/api.dart';
+import 'package:flup_openapi_lib/api.dart';
 
 
 final api_instance = ExpensesApi();
