+++ frontend/flup_openapi_lib/lib/api/expenses_api.dart
