+++ frontend/flup_openapi_lib/lib/api_exception.dart
