+++ frontend/flup_openapi_lib/lib/auth/http_basic_auth.dart
