+++ frontend/flup_openapi_lib/lib/auth/http_bearer_auth.dart
