+++ frontend/flup_openapi_lib/lib/model/expense.dart
