+++ frontend/flup_openapi_lib/pubspec.yaml
@@ -2,7 +2,7 @@
 # AUTO-GENERATED FILE, DO NOT MODIFY!
 #
 
-name: 'test_api'
+name: 'flup_openapi_lib'
 version: '1.0.0'
 description: 'OpenAPI API client'
 homepage: 'homepage'
