+++ frontend/generate_flup_openapi_lib/README.md
@@ -1,0 +1,2 @@
+Usage:
+`flutter pub run build_runner build --delete-conflicting-outputs`
