+++ frontend/generate_flup_openapi_lib/lib/generate_flup_openapi_lib.dart
@@ -1,0 +1,9 @@
+import 'package:openapi_generator_annotations/openapi_generator_annotations.dart';
+
+@Openapi(
+    additionalProperties:
+    AdditionalProperties(pubName: 'flup_openapi_lib', pubAuthor: 'Laura'),
+    inputSpecFile: '../../backend/swagger/swagger.json',
+    generatorName: Generator.dart,
+    outputDirectory: '../flup_openapi_lib')
+class Example extends OpenapiGeneratorConfig {}
